+++ src/corner.h
@@ -9,21 +9,21 @@ using namespace std;
 class Corner
 {
 private:
-	vector<vector<long> > permutation_transition_table, orientation_transition_table;
+	vector<vector<long long> > permutation_transition_table, orientation_transition_table;
 	vector<char> prune_table;
 	int state_count;
 	string prune_file_path;
 
 public:
-	Corner(vector<vector<long> > & permutation_transition_table, vector<vector<long> > & orientation_transition_table)
+	Corner(vector<vector<long long> > & permutation_transition_table, vector<vector<long long> > & orientation_transition_table)
 		: permutation_transition_table(permutation_transition_table), orientation_transition_table(orientation_transition_table), 
 		  state_count(88179840), prune_file_path("corner_prune.data") {}
 	void buildPruneTable();
 	vector<char> getPruneTable(){ return prune_table; }
 private:
-	void pruneTreeSearch(long state, char depth_left, char depth, int lastMove);
-	long apply_transition(long state, int move);
-	bool solveable(long state, char depth, char maxBreathDepthSearch, int lastMove);
+	void pruneTreeSearch(long long state, char depth_left, char depth, int lastMove);
+	long long apply_transition(long long state, int move);
+	bool solveable(long long state, char depth, char maxBreathDepthSearch, int lastMove);
 	void store_to_file_prune(string path);
 	bool try_read_from_file(string path);
 };
